[BOOK-170] style: swift lint
