+++ src/Projects/BKPresentation/Sources/ArchiveFlow/Model/BookStatus.swift
@@ -1,0 +1,10 @@
+// Copyright © 2025 Booket. All rights reserved
+
+import Foundation
+
+enum BookStatus: String, CaseIterable {
+    case total = "전체"
+    case toRead = "읽기 전"
+    case reading = "읽는 중"
+    case completed = "완독"
+}
