+++ src/Projects/BKPresentation/Sources/ArchiveFlow/Model/ChipData.swift
@@ -11,7 +11,35 @@ struct ChipData: Hashable, Equatable {
         hasher.combine(title)
     }
     
-    static func == (lhs: ChipData, rhs: ChipData) -> Bool {
+    static func == (
+        lhs: ChipData,
+        rhs: ChipData
+    ) -> Bool {
         lhs.title == rhs.title
+    }
+}
+
+enum ChipType: Int, CaseIterable {
+    case total = 0
+    case toRead = 1
+    case reading = 2
+    case completed = 3
+    
+    var title: String {
+        switch self {
+        case .total: return "전체"
+        case .toRead: return "읽기 전"
+        case .reading: return "읽는 중"
+        case .completed: return "완독"
+        }
+    }
+    
+    var bookStatus: BookStatus? {
+        switch self {
+        case .total: return .total
+        case .toRead: return .toRead
+        case .reading: return .reading
+        case .completed: return .completed
+        }
     }
 }
